[Turbopack] only invalidate fs writes with a different content (#75936)
### What?

There are cases where two write tasks write to the same file. This can happen e. g. when two different image source files with the same content are used and end up with the same content hashed filename.

But since dependency tracking is disabled for one off builds, this results in an invalidation of the first write task when the second write task effect is executed. And an invalidation will panic in dependency-tracking-disabled mode.

This change only invalidates other write tasks when they write a different content and keeps writes with the same content intact.

This should avoid the panic as no task is invalidated.