[dev-overlay] deprecate `devIndicators.buildActivityPosition` and rename to `position` (#76069)
### Why?

Since the `buildActivity` is deprecated at #76067, `buildActivityPosition` may also be deprecated. However, we are planning to preserve the option to position the dev indicator. Therefore, this PR adds a new option `position` and deprecates and aliases the `buildActivityPosition` to `position`.

### Success Criteria

Deprecation:

- [x] Does using `buildActivityPosition` show a deprecation warning?
- [x] Does `buildActivityPosition` have a deprecation in JSDoc?

Alias:

> Is it backward compatible with the old overlay?

- [x] Does setting `buildActivityPosition` correctly set the build activity indicator position?
- [x] Does it warn if `buildActivityPosition` is set instead of `position`?
- [x] Does `buildActivityPosition` supersede `position` value to prevent breaking change?
- [x] Does setting the `position` correctly set the build activity indicator position?

### Follow Up

In the following PRs, we will add the behavior for `position` to adjust the dev indicator's position. After removing the old overlay's sources, we can remove any logic that `buildActivityPosition` was holding for the old overlay and make it fully an alias for backwards compatibility.

Closes NDX-837