Fix runtime error in `writeConfigurationDefaults` (#76106)
Follow-up for #74674

When running `pnpm next build` or `pnpm next dev` with a test app in the
Next.js repo, we patch the app's `tsconfig.json` to exclude the test
files from the program, because those are already covered by the root
`tsconfig.json`.

The code to suppress logging this action had a bug where we tried to
remove the action from the list of `requiredActions` even if it wasn't
added, resulting in the following runtime error:

```
Cannot read properties of undefined (reading 'includes')
```

This could happen when the `requiredActions` array was empty but the
`suggestedActions` array was not.