[Segment Cache] Implement scroll-to-new-page
This implements Next's scroll-to-new-page behavior the Segment Cache.

The LayoutRouter expects an array of segment paths that represents
every page in the new route tree. In the old implementation, this is
generated by the `generateSegmentsFromPatch` function. In the new
implementation, I've chosen to generate this array during the main
tree traversal that happens in `startPPRNavigation`.

As a future improvement, we should use the same segment path type that
we use to access Segment Cache entries; there's no reason to have two
representations of the same concept.