[dev-overlay]: make open in editor button a separate click region (#76243)
Currently the code frame/terminal headers are treated as a clickable
region to open in your editor. However, the editor integration might not
always work or be desired, and folks might instead want to copy the file
path for other reasons.

This moves the "open in editor" functionality to be a specific click
area around the icon, removing interactivity from the header & call
stack frames.



https://github.com/user-attachments/assets/e7e1c282-233f-4ca1-9668-f520f0b7bbe8



Closes NDX-794
Closes NDX-802

---------

Co-authored-by: devjiwonchoi <devjiwonchoi@gmail.com>