[dev-overlay] use unique font name (#76230)
We [inject a style tag](https://github.com/vercel/next.js/pull/76225)
that downloads Geist fonts from local file system. However, that style
tag is injected to the main DOM because of an
[issue](https://issues.chromium.org/issues/41085401) in Chrome that the
fonts do not work in shadow DOM. We worry that this injected style would
interfere with user's styles, or vice versa. Hence, we are going to name
them with unique prefix to avoid most collision.