[app i18n]: warn about i18n configuration deprecation in app router (#76346)
The `i18n` configuration was supported in app router to assist with
incremental migration, but it's only meant to be used for pages router.
There are some bugs that can pop up when attempting to use it with app
router.

Disabling `i18n` configuration entirely for any app router paths will be
handled in a separate PR: the goal for this PR is to signal that it's
unsupported in app router.