refactor: clean up forked Form implementations (#76495)
Split out of #76184 because it got a bit involved and i want to do it in
isolation.

Removes dead code from the forked implementations of Form for app and
pages dir. There's some shared utils but not much. Mostly involves
removing code for prefetching and server actions from the pages
implementation.

Also took this opportunity to fix some stuff for pages:
- improve error messages, previously they were linking to `/app/` docs
- use `React.forwardRef`: technically, you can use React <19 in pages,
where `forwardRef` is still needed