Run React 18 tests for automated update PRs
This ensures that we don't marks tests as `passed` in the bundler test
manifests if they would still fail when running them with React 18.

Those tests are usually only run on `canary`, unless a PR has the
`run-react-18-tests` label, which we are now automatically adding for
the update PRs.