Webpack: abstract away getting compilation spans
Currently many of our plugins import a singleton mapping of webpack compilation objects to their corresponding tracing span.

This change abstracts that away into a function called `getCompilationSpan`, and allows us to use the corresponding mapping in the rspack case.

Previously, any plugin that used these would fail with rspack when the compilation object wouldn’t be found. This was the case when running `next build` with rspack.

Test Plan: Install and run `next build` with the rspack plugin in a test app.