Unflake `parallel-routes-revalidation` test
[flakiness metric](https://app.datadoghq.com/ci/test/runs?query=test_level%3Atest%20env%3Aci%20%40git.repository.id%3Agithub.com%2Fvercel%2Fnext.js%20%40test.service%3Anextjs%20%40test.status%3Afail%20%28-%40git.branch%3A%2A%3F%2A%20OR%20%40git.branch%3Acanary%29%20%40test.name%3A%22parallel-routes-revalidation%20should%20not%20trigger%20the%20intercepted%20route%20when%20lazy-fetching%20missing%20data%22&agg_m=count&agg_m_source=base&agg_t=count&citest_explorer_sort=timestamp%2Casc&cols=%40test.status%2Ctimestamp%2C%40test.suite%3A83%2C%40test.name%3A607%2C%40git.branch&currentTab=overview&eventStack=&fromUser=false&graphType=flamegraph&index=citest&start=1738068572273&end=1740660572273&paused=false)

The latest canary release
[failed](https://github.com/vercel/next.js/actions/runs/13555525831/job/37890389323)
because the deploy test `parallel-routes-revalidation › should not
trigger the intercepted route when lazy-fetching missing data` kept
failing even with retries.

According to my local testing using `pnpm test-deploy
test/e2e/app-dir/parallel-routes-revalidation/parallel-routes-revalidation.test.ts
-t 'should not trigger the intercepted route when lazy-fetching missing
data'`, the forward navigation was sometimes triggered too early after
the page was reloaded.

This PR introduces an option to wait for hydration (same as we do for
the initial load) when reloading a page, which fixed the test for me.

Note: I don't like the option name `waitHydration`, but went with it
anyways because we're already using it for the `next.browser` options.