Revert "Run nightly bundler integration tests also with React 18" (#76640)
Reverts vercel/next.js#76606

Causes the nightly jobs to fail with:

```
Error: Failed to CreateArtifact: Received non-retryable error: Failed request: (409) Conflict: an artifact with this name already exists on the workflow run
```

because we're running two `build_reusable` jobs with the same stepName.
While I could fix this, we'd still have problems with either duplicating
tests on areweturboyet, or we'd need some logic to merge test results
across the matrix.

I like the idea of running the full matrix, but this type of thing is
rare and maybe not worth solving for (aside from catching it when the PR
goes up, which we now have logic for, thanks to @unstubbable).