[dynamicIO] routes with dynamic segments should be able to be static in dev (#76691)
Dev mode reports static/dynamic behavior using a heuristic when
dynamicIO is enabled. The heuristic previously incorrectly marked routes
as dynamic when they were in fact static as a consequence of refactoring
metadata to support streaming. This update adjust the heuristic to not
incorrectly assume dynamic in these situations.

The root cause is that errors in `onError` cannot be assumed to be valid
if the render has already aborted

Closes NAR-112