[dev-overlay] Always show relative paths
Stackframes that are not sourcemapped used to show
absolute paths when the input was also an absolute path.
This looks off when it is mixed with frames that
sourcemapped and we display relative paths.

Since we also have an explicit indicator for failed
sourcemapping, we don't need to keep absolute file
paths as a proxy.