fix edge runtime asset fetch in pages api (#76750)
### What

We discoved an issue that fetch new URL with static assets in Pages API on edge runtime were failing. Located the issue was an introduced in #66534 where we changed the layer of Pages API with edge runtime entries into API layer, where it actually should be another separate layer like middleware.

The quick fix is using middleware bundle layer for Pages API edge functions, but middleware is server-only layer which Pages API is not yet, since it doesn't bundle everything and picking up the `react-server` condition. So we have to create a new layer `apiEdge` to minimize the behavior changes and also fix the bug.

The test suite was disabled due to node-fetch blocker, but we now we can enable it.

Closes NDX-941
Fixes #74385 
[slack-ref](https://vercel.slack.com/archives/C07CGAA2RS6/p1740868643335099)